+++ platform/apps/grafana/terraform/locals.tf
@@ -1,9 +1,13 @@
 locals {
   slug = "grafana"
 
-  authorized_groups = var.enabled ? merge(
-    { (var.access_level) = var.base.authentik.groups[var.access_level] },
-    var.extra_authorized_groups,
+  authorized_groups = var.enabled ? (
+    var.tier_cascade_enabled
+    ? var.base.authentik.tier_cascade[var.tier_access_level]
+    : merge(
+      { (var.access_level) = var.base.authentik.groups[var.access_level] },
+      var.extra_authorized_groups,
+    )
   ) : {}
 
   app_url           = "https://${var.hostname}"